Time: 0 ms (100.00%) | Memory: 7 MB (80.96%) - LeetSync

+++ 910-nth-magical-number/nth-magical-number.cpp
@@ -1,7 +1,7 @@
 const int MOD = 1e9 + 7;
 class Solution {
 public:
-    bool isPossible(long long mid, long long n, int a, int b) {
+    bool isPossible(long long mid, int n, int a, int b) {
         long long x = mid / a;
         long long y = mid / b;
         long long z = mid / lcm(a, b);
